+++ app/Http/Controllers/Admin/VehicleBookingController.php
@@ -9,6 +9,7 @@ use App\Models\Vehicle;
 use App\Models\CrewProfile;
 use App\Models\Customer;
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
 use App\Models\Payment;
 use Carbon\Carbon;
 use Maatwebsite\Excel\Facades\Excel;
@@ -29,6 +30,7 @@ class VehicleBookingController extends Controller
      */
     public function index(Request $request)
     {
+        Gate::authorize('index_vehicles_vehicle_bookings');
         $query = VehicleBooking::with([
             'vehicle',
             'customer',
@@ -69,6 +71,7 @@ class VehicleBookingController extends Controller
      */
     public function create(Request $request)
     {
+        Gate::authorize('create_vehicles_vehicle_bookings');
         $vehicles = Vehicle::all();
         $drivers = CrewProfile::whereHas('user', function ($q) {
             $q->where('role', 'driver');
@@ -88,6 +91,7 @@ class VehicleBookingController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('create_vehicles_vehicle_bookings');
         $request->validate([
             'vehicle_id' => 'required',
             'customer_id' => 'required|exists:customers,id',
@@ -152,6 +156,7 @@ class VehicleBookingController extends Controller
 
     public function edit(VehicleBooking $vehicleBooking)
     {
+        Gate::authorize('update_vehicles_vehicle_bookings');
         $vehicles = Vehicle::all();
         $drivers = CrewProfile::whereHas('user', function ($q) {
             $q->where('role', 'driver');
@@ -186,6 +191,8 @@ class VehicleBookingController extends Controller
 
     public function update(Request $request, VehicleBooking $vehicleBooking)
     {
+        Gate::authorize('update_vehicles_vehicle_bookings');
+        
         $updateData = $request->all();
         $updateData['start_time'] = $request->start_time;
         $updateData['end_time'] = $request->end_time;
@@ -228,6 +235,7 @@ class VehicleBookingController extends Controller
 
     public function show(VehicleBooking $vehicleBooking)
     {
+        Gate::authorize('view_vehicles_vehicle_bookings');
         $vehicleBooking->load([
             'vehicle',
             'customer',
@@ -243,6 +251,7 @@ class VehicleBookingController extends Controller
 
     public function destroy(VehicleBooking $vehicleBooking)
     {
+        Gate::authorize('delete_vehicles_vehicle_bookings');
         try {
             $vehicleBooking->delete();
